idest/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0

bcrypt==4.0.1

cffi==1.15.1

django-filter==22.1
django-taggit==3.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
Faker==15.3.2
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2==2.9.7
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0

whitenoise==6.3.0
wincertstore==0.2

